Test for PluginSpecEditor
